+++ ESP32/Smart_Home_Energy_Management_System/code/lib/Display/led_indicator/led_indicator.h
@@ -2,6 +2,6 @@
 
 namespace led_indicator
 {
-  void begin(uint8_t pin);
+  void begin(int pin);
   void update();
 }
